+++ Sources/Sentry/SentryViewHierarchyIntegration.m
@@ -40,6 +40,8 @@ saveViewHierarchy(const char *reportDirectoryPath)
 
     sentrycrash_setSaveViewHierarchy(&saveViewHierarchy);
 
+    SentryDependencyContainer.sharedInstance.viewHierarchy.reportAccessibilityIdentifier
+        = options.reportAccessibilityIdentifier;
     return YES;
 }